+++ Dockerfile
@@ -10,6 +10,7 @@ RUN crontab /etc/cron.d/ml-work-cronjob
 
 COPY note/train_img /code/train_img/
 COPY src/mnist/main.py /code/
+COPY src/mnist/model/ /code/
 COPY run.sh /code/run.sh
 
 RUN chmod +x /code/run.sh
